modules/CloudFunction/src/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.16.0
Flask==2.3.3  